+++ iosApp/iosApp.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,0 +1,61 @@
+{
+  "object": {
+    "pins": [
+      {
+        "package": "BridgeArchiver",
+        "repositoryURL": "https://github.com/Sage-Bionetworks/BridgeArchiver-Swift.git",
+        "state": {
+          "branch": null,
+          "revision": "5dc7b0d00d568fbdb4a32993bc18bd55c82a9de4",
+          "version": "0.1.0"
+        }
+      },
+      {
+        "package": "CMSSupport",
+        "repositoryURL": "https://github.com/Sage-Bionetworks/CMSSupport.git",
+        "state": {
+          "branch": null,
+          "revision": "7390eb96b9edf53c3127d2aa2641a9475b36136e",
+          "version": "1.2.0"
+        }
+      },
+      {
+        "package": "JsonModel",
+        "repositoryURL": "https://github.com/Sage-Bionetworks/JsonModel-Swift.git",
+        "state": {
+          "branch": null,
+          "revision": "d45397f95a7ca9368efbf94ab1ab720edf0a08a9",
+          "version": "1.2.1"
+        }
+      },
+      {
+        "package": "MobilePassiveData",
+        "repositoryURL": "https://github.com/Sage-Bionetworks/MobilePassiveData-SDK.git",
+        "state": {
+          "branch": null,
+          "revision": "6b8b0205e226ec25960f1369db107665334e0d34",
+          "version": "1.0.0"
+        }
+      },
+      {
+        "package": "SageResearch",
+        "repositoryURL": "https://github.com/Sage-Bionetworks/SageResearch.git",
+        "state": {
+          "branch": null,
+          "revision": "9a639da0b74d3fb81c20d107fd5b5dbfb1dd2fdd",
+          "version": "4.1.4"
+        }
+      },
+      {
+        "package": "ZIPFoundation",
+        "repositoryURL": "https://github.com/weichsel/ZIPFoundation.git",
+        "state": {
+          "branch": null,
+          "revision": "cf10bbff6ac3b873e97b36b9784c79866a051a8e",
+          "version": "0.9.12"
+        }
+      }
+    ]
+  },
+  "version": 1
+}
